# Conflicts:
#	CMakeLists.txt
#	Expressions/Command/DefineVarCommand.cpp
#	Expressions/Command/PrintCommand.cpp
#	main.cpp

+++ SymbolTable.h
@@ -15,7 +15,8 @@ private:
 
     map<string, string> symbolToPathMap;
 
-    map<string, int> pathToValueIndexMap;
+    map<string, double > pathToValueIndexMap;
+
     double values[VALUES_SIZE] = {0};
 
     //table of non binded values
@@ -23,9 +24,8 @@ private:
 
 
     SymbolsTable();
+
     double getValueFromPath(string path);
-
-
 
 
 public:
@@ -33,14 +33,23 @@ public:
     static SymbolsTable *getInstance();
 
     void addSymbol(string symbol, string path);
+
     bool exist(string symbol);
+
     virtual ~SymbolsTable();
+
     void updateValues(string data);
+
     double getCommonValue(string symbol);
+
     string getPath(string symbol);
+
     void addTempValue(string symbol, double value);
+
     bool isTempValue(string symbol);
+
     double getTempValue(string symbol);
+
     bool isCommonSymbol(string symbol);
 
 };
